Added some missing tests
