+++ internal/feature/article/src/test/kotlin/com/recco/internal/feature/article/Stubs.kt
@@ -22,18 +22,6 @@ internal fun RecommendationRepository.stubForSuccessWithLikedAndBookmarkedArticl
     stubRepositoryForSuccess(createArticle(rating = Rating.LIKE))
 }
 
-internal fun RecommendationRepository.stubForSuccessWithDislikedArticle() {
-    stubRepositoryForSuccess(createArticle(rating = Rating.DISLIKE))
-}
-
-internal fun RecommendationRepository.stubForSuccessWithNonRatedArticle() {
-    stubRepositoryForSuccess(createArticle())
-}
-
-internal fun RecommendationRepository.stubForSuccessWithNonBookmarkedArticle() {
-    stubRepositoryForSuccess(createArticle(isBookmarked = false))
-}
-
 internal fun RecommendationRepository.stubForInitialFailure() {
     stub {
         onBlocking { it.getArticle(any()) } doThrow staticThrowableForTesting
@@ -54,31 +42,10 @@ internal fun RecommendationRepository.stubForToggleRatingFailure() {
     }
 }
 
-internal fun RecommendationRepository.stubForToggleBookmarkFailure() {
-    stub {
-        onBlocking { it.setRecommendationAsViewed(any()) } doReturn Unit
-        onBlocking { it.getArticle(any()) } doReturn createArticle(rating = Rating.LIKE)
-        onBlocking {
-            it.setBookmarkRecommendation(
-                any(),
-                any()
-            )
-        } doThrow staticThrowableForTesting
-    }
-}
-
 private fun RecommendationRepository.stubRepositoryForSuccess(article: Article) {
     stub {
         onBlocking { it.setRecommendationAsViewed(any()) } doReturn Unit
         onBlocking { it.getArticle(any()) } doReturn article
-        onBlocking { it.setBookmarkRecommendation(any(), any()) } doReturn Unit
-        onBlocking { it.setRecommendationRating(any(), any()) } doReturn Unit
-    }
-}
-
-private fun RecommendationRepository.stubRepositoryForSuccess(id: ContentId) {
-    stub {
-        onBlocking { it.setRecommendationAsViewed(any()) } doReturn Unit
         onBlocking { it.setBookmarkRecommendation(any(), any()) } doReturn Unit
         onBlocking { it.setRecommendationRating(any(), any()) } doReturn Unit
     }
