+++ internal/feature/rating/src/test/kotlin/ContentInteractViewModelDelegateTest.kt
@@ -5,6 +5,7 @@ import com.recco.internal.core.model.recommendation.Rating
 import com.recco.internal.core.model.recommendation.UserInteractionRecommendation
 import com.recco.internal.core.repository.RecommendationRepository
 import com.recco.internal.core.test.CoroutineTestExtension
+import com.recco.internal.core.test.utils.staticThrowableForTesting
 import com.recco.internal.feature.rating.delegates.ContentInteractViewModelDelegate
 import com.recco.internal.feature.rating.delegates.ContentUserInteract
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -15,8 +16,11 @@ import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.extension.ExtendWith
 import org.mockito.kotlin.any
 import org.mockito.kotlin.doReturn
+import org.mockito.kotlin.doThrow
 import org.mockito.kotlin.mock
 import org.mockito.kotlin.stub
+import org.mockito.kotlin.times
+import org.mockito.kotlin.verifyBlocking
 
 @OptIn(ExperimentalCoroutinesApi::class)
 @ExtendWith(CoroutineTestExtension::class)
@@ -230,7 +234,51 @@ class ContentInteractViewModelDelegateTest {
             )
     }
 
+    @Test
+    fun `onFailure emits exceptions while logging them if ToggleDislikeState`() = runTest {
+        val viewModelDelegate = createViewModelDelegate(
+            startWith = reccomendationAs(
+                isDislikeLoading = true,
+                isLikeLoading = false,
+                rating = Rating.NOT_RATED,
+            )
+        )
 
+        repository.stubForToggleRatingFailure()
+
+        viewModelDelegate.onContentUserInteract(
+            ContentUserInteract.ToggleDislikeState(id)
+        )
+
+        advanceUntilIdle()
+
+        verifyBlocking(logger, times(1)) {
+            e(staticThrowableForTesting, null, null)
+        }
+    }
+
+    @Test
+    fun `onFailure emits exceptions while logging them if ToggleBookmarkState`() = runTest {
+        val viewModelDelegate = createViewModelDelegate(
+            startWith = reccomendationAs(
+                isDislikeLoading = true,
+                isLikeLoading = false,
+                rating = Rating.NOT_RATED,
+            )
+        )
+
+        repository.stubForToggleBookmarkFailure()
+
+        viewModelDelegate.onContentUserInteract(
+            ContentUserInteract.ToggleBookmarkState(id)
+        )
+
+        advanceUntilIdle()
+
+        verifyBlocking(logger, times(1)) {
+            e(staticThrowableForTesting, null, null)
+        }
+    }
 
 
     private fun createViewModelDelegate(startWith: UserInteractionRecommendation): ContentInteractViewModelDelegate {
@@ -242,6 +290,18 @@ class ContentInteractViewModelDelegateTest {
         }
     }
 
+    internal fun RecommendationRepository.stubForToggleBookmarkFailure() {
+        stub {
+            onBlocking { it.setRecommendationAsViewed(any()) } doReturn Unit
+            onBlocking {
+                it.setBookmarkRecommendation(
+                    any(),
+                    any()
+                )
+            } doThrow staticThrowableForTesting
+        }
+    }
+
     private fun RecommendationRepository.stubRepositoryForSuccess(id: ContentId) {
         stub {
             onBlocking { it.setRecommendationAsViewed(any()) } doReturn Unit
@@ -249,6 +309,19 @@ class ContentInteractViewModelDelegateTest {
             onBlocking { it.setRecommendationRating(any(), any()) } doReturn Unit
         }
     }
+
+    private fun RecommendationRepository.stubForToggleRatingFailure() {
+        stub {
+            onBlocking { it.setRecommendationAsViewed(any()) } doReturn Unit
+            onBlocking {
+                it.setRecommendationRating(
+                    any(),
+                    any()
+                )
+            } doThrow staticThrowableForTesting
+        }
+    }
+
 
     private fun reccomendationAs(
         id: ContentId = this.id,
